backend/registration_agent/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.1.0
openai==1.81.0
pydantic==2.11.4
pyairtable==2.3.3
gocardless-pro==3.1.0
email-validator==2.1.0
boto3==1.35.90
twilio==8.5.0
# HEIC image conversion support
pillow==10.1.0
pillow-heif==0.17.0 